Split svd function defs into new file svd.cc
